fix: align schema version with AGENTS.md v1.2.0
